Refactor cache mod, remove linked_hash_map
