+++ ballista/cache/Cargo.toml
@@ -26,6 +26,7 @@ edition = "2021"
 async-trait = "0.1.64"
 futures = "0.3"
 hashbrown = "0.14"
+hashlink = "0.8.4"
 log = "0.4"
 parking_lot = "0.12"
 tokio = { version = "1.25", features = ["macros", "parking_lot", "rt-multi-thread", "sync", "time"] }
